requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
ultralytics==8.3.10
pillow==10.4.0
torch>=2.1.0
